+++ src/main/java/com/aiwan/scenes/Dao/ScenesDao.java
@@ -6,5 +6,5 @@ import com.aiwan.scenes.protocol.CM_Shift;
 public interface ScenesDao {
     public void updateUserPosition(CM_Move cm_move);
 
-    public void updateMapPosition(CM_Shift cm_shift,short x,short y);
+    public void updateMapPosition(CM_Shift cm_shift,int x,int y);
 }
